Return decimal instead of float

+++ cryex/core.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 import requests
+from decimal import Decimal
 
 
 class ExchangeError(Exception):
@@ -36,11 +37,11 @@ class Poloniex(Exchange):
 
         return {
             'exchange': 'poloniex',
-            'last': float(data['last']),
+            'last': Decimal(data['last']),
             'pair': pair,
-            'volume24h': float(data['quoteVolume']),
-            'high24h': float(data['high24hr']),
-            'low24h': float(data['low24hr']),
+            'volume24h': Decimal(data['quoteVolume']),
+            'high24h': Decimal(data['high24hr']),
+            'low24h': Decimal(data['low24hr']),
         }
 
 
@@ -65,9 +66,9 @@ class Kraken(Exchange):
         data = requests.get(url).json()['result'][new_pair]
         return {
             'exchange': 'kraken',
-            'last': float(data['c'][0]),
+            'last': Decimal(data['c'][0]),
             'pair': pair,
-            'volume24h': float(data['v'][1]),
-            'high24h': float(data['h'][1]),
-            'low24h': float(data['l'][1]),
+            'volume24h': Decimal(data['v'][1]),
+            'high24h': Decimal(data['h'][1]),
+            'low24h': Decimal(data['l'][1]),
         }
